requirements.txt
groq==0.11.0
python-docx==1.1.2
httpx==0.27.0
